client: gearman client now faster and better
Eliminated the strange usage of in favor of using channels. Removed the
redundant `processLoop` goroutine, the `readLoop` now inlines the
processing (it was serial anyway!). Switched to the use of threadsafe
memory object pools for request and response objects, and now allows
pipelining if multiple goroutines are making requests at once.

Removed a lot of redundancy in the code, and eliminated the use of the
strange `responseHandlerMap` in favor of a concurrent map for holding
callbacks on a per-handler basis.

Refactored the `request` type such that it now uses a builder pattern,
and now itself keeps a small byte buffer for each request which is
reused when parameters are attached to requests. Streamlined the
formatting of outbound messages such that they are encoded as they are
written to the output buffer, to reduce the number of copies involved in
most operations.

Eliminated the strange blind-read loop which pretended it had no idea
how large of a buffer it would need to allocate to read a response, and
now allocates a buffer for each payload, the size of the payload as
specified by the server, respectively.