client: Improved error handling, added snapshot test
Added a simple tool to the client test to replay snapshots captured by
`tcpflow` so long as those snapshots have a manifest indicating what the
ip and port of the sending and receiving processes were. This can be
used to capture any conversation with gearman and replay it; the tool
should probably be factored out into a common set of TCP protocol
testing utilities.

Captured a client/server conversation with a the perl `Gearman::Client`
and `Gearman::Worker` APIs against a C++ `gearmand` and verified that
the client handles the replay as expected.

Eliminated the remaining `sync.Mutex` from the client library in favor
of using compare-and-swap to access the pointer to the connection (which
is exposed as an interface value type). It turns out that interface
value types in Go are a pair of pointers represented typically as a 128
bit value, and are immutable from within the programming language. As a
simple work-around, a pointer to a struct which includes that interface
through composition was used in its stead, and CAS was used on that
pointer. This allowed the pointer to generally be used as-is without
having to be derferenced or having to deal with a pointer to an
interface (which has two levels of possibly-nil values).

Added the ability for the server stream to return an `ERROR` message as
a response to `SUBMIT_JOB_*` which in practice is what `gearmand` will
do when it was not able to succcessfully enqueue the job. This cleans up
the behavior of a client if one of the requests it makes does yield an
error.

Further, removed the additional level of buffering on `net.Conn`
connections; under the covers this is buffered non-blocking I/O anyway
so an additional `bufio` buffer was completely unnecessarry.