Add a sense of ownership of which thread (read or write) is reconnecting
Problem: suppose readLoop() encountered an error on the connection,
it starts the reconnect process, but suppose at nearly the same time the
writeLoop() thread was using the same connection handle and also encountered
an error (since connection is broken). This fix is to prevent both of them attempting
to reconnect and have one of them claim the ownership of reconnect process
and have the other thread merely wait for it to complete and exit reconnect,
and then go on with using the new conn handle via getConn().